Using sqrtf() in CCLayer.m to prevent a compiler warning
Former-commit-id: 97ebbbb284ce531e08646ef59b25423830f3895d

+++ cocos2d/CCLayer.m
@@ -456,7 +456,7 @@
     if (h == 0)
 		return;
 
-	double c = sqrt(2);
+	float c = sqrtf(2);
     CGPoint u = ccp(vector_.x / h, vector_.y / h);
 
 	// Compressed Interpolation mode
